+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.0.0-alpha.12](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.11...1.0.0-alpha.12) (2021-09-21)
+
+
+### Features
+
+* register device token with FCM ([#46](https://github.com/customerio/customerio-ios/issues/46)) ([f6a87e0](https://github.com/customerio/customerio-ios/commit/f6a87e0e889b9fdac2c8093a8f4fb4f5f0e9ea2a))
+
+# [1.0.0-alpha.11](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.10...1.0.0-alpha.11) (2021-09-20)
+
+
+### Features
+
+* event tracking ([#42](https://github.com/customerio/customerio-ios/issues/42)) ([be768bb](https://github.com/customerio/customerio-ios/commit/be768bbb32680ac0ca8dce098f62720394965f85))
+
+# [1.0.0-alpha.10](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.9...1.0.0-alpha.10) (2021-09-15)
+
+
+### Bug Fixes
+
+* remove apn device token profile request body ([#41](https://github.com/customerio/customerio-ios/issues/41)) ([61946c4](https://github.com/customerio/customerio-ios/commit/61946c4dea617561891188aa537277cc67af35ae))
+
 # [1.0.0-alpha.9](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.8...1.0.0-alpha.9) (2021-09-15)